+++ java/src/class2/exercice5/abstracts/BankAccount.java
@@ -5,13 +5,13 @@ import java.util.List;
 import class2.exercice5.Movement;
 
 public abstract class BankAccount {
-    private final Integer id;
+    private final Long id;
     protected Double balance;
     private final boolean special;
     private final Type type;
     protected List<Movement> movements;
 
-    public BankAccount(final Integer id, Double initialBalance, Type type, boolean special) {
+    public BankAccount(final Long id, Double initialBalance, Type type, boolean special) {
         this.id = id;
         this.balance = initialBalance;
         this.type = type;
@@ -39,7 +39,7 @@ public abstract class BankAccount {
         return new ArrayList<>(this.movements);
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
